refactor: golraba-spring-api와 db-core 순환 참조하지 않도록 변경
