+++ build.gradle
@@ -1,17 +1,14 @@
 plugins {
-    id 'java'
+    id 'java-library'
     id 'org.springframework.boot' apply false
     id 'io.spring.dependency-management'
     id 'org.asciidoctor.jvm.convert' apply false
 }
 
-java {
-    sourceCompatibility = '17'
-}
-
 allprojects {
     group = "${projectGroup}"
     version = "${applicationVersion}"
+    sourceCompatibility = project.javaVersion
 
     repositories {
         mavenCentral()
@@ -19,7 +16,7 @@ allprojects {
 }
 
 subprojects {
-    apply plugin: 'java'
+    apply plugin: 'java-library'
     apply plugin: 'org.springframework.boot'
     apply plugin: 'io.spring.dependency-management'
     apply plugin: 'org.asciidoctor.jvm.convert'
