+++ golraba-spring-api/src/main/java/donggi/dev/kkeuroolryo/core/comment/application/CommentService.java
@@ -4,8 +4,8 @@ import donggi.dev.kkeuroolryo.core.comment.application.dto.CommentDto;
 import donggi.dev.kkeuroolryo.core.comment.application.dto.CommentPaginationDto;
 import donggi.dev.kkeuroolryo.core.comment.domain.Comment;
 import donggi.dev.kkeuroolryo.core.comment.domain.CommentRepository;
-import donggi.dev.kkeuroolryo.core.comment.domain.exception.CommentNotFoundException;
-import donggi.dev.kkeuroolryo.core.comment.domain.exception.NoOffsetPageInvalidException;
+import donggi.dev.kkeuroolryo.core.comment.exception.CommentNotFoundException;
+import donggi.dev.kkeuroolryo.core.comment.exception.NoOffsetPageInvalidException;
 import donggi.dev.kkeuroolryo.core.question.domain.QuestionRepository;
 import donggi.dev.kkeuroolryo.core.question.domain.exception.QuestionNotFoundException;
 import donggi.dev.kkeuroolryo.web.comment.dto.CommentRegisterCommand;
