+++ golraba-spring-api/src/main/java/donggi/dev/kkeuroolryo/core/comment/domain/CommentPassword.java
@@ -1,6 +1,6 @@
 package donggi.dev.kkeuroolryo.core.comment.domain;
 
-import donggi.dev.kkeuroolryo.core.comment.domain.exception.CommentInvalidPasswordException;
+import donggi.dev.kkeuroolryo.core.comment.exception.CommentInvalidPasswordException;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
 import java.util.Objects;
