+++ golraba-spring-api/src/main/java/donggi/dev/kkeuroolryo/core/comment/exception/CommentUnauthorizedException.java
@@ -1,4 +1,4 @@
-package donggi.dev.kkeuroolryo.core.comment.domain.exception;
+package donggi.dev.kkeuroolryo.core.comment.exception;
 
 import donggi.dev.kkeuroolryo.common.exception.ErrorCodeAndMessage;
 import donggi.dev.kkeuroolryo.common.exception.GolrabaException;
