+++ golraba-spring-api/src/test/java/donggi/dev/kkeuroolryo/core/comment/application/CommentEditorTest.java
@@ -7,8 +7,8 @@ import donggi.dev.kkeuroolryo.IntegrationTest;
 import donggi.dev.kkeuroolryo.core.comment.application.dto.CommentDto;
 import donggi.dev.kkeuroolryo.core.comment.domain.Comment;
 import donggi.dev.kkeuroolryo.core.comment.domain.CommentRepository;
-import donggi.dev.kkeuroolryo.core.comment.domain.exception.CommentNotFoundException;
-import donggi.dev.kkeuroolryo.core.comment.domain.exception.CommentUnauthorizedException;
+import donggi.dev.kkeuroolryo.core.comment.exception.CommentNotFoundException;
+import donggi.dev.kkeuroolryo.core.comment.exception.CommentUnauthorizedException;
 import donggi.dev.kkeuroolryo.core.question.domain.Question;
 import donggi.dev.kkeuroolryo.core.question.domain.QuestionRepository;
 import donggi.dev.kkeuroolryo.core.question.domain.exception.QuestionNotFoundException;
